combine localstorage_token.js into accounts_client.js
