+++ packages/accounts-base/accounts_client.js
@@ -177,6 +177,437 @@ export class AccountsClient extends AccountsCommon {
       }
     );
   }
+
+  ///
+  /// LOGIN METHODS
+  ///
+
+  // Call a login method on the server.
+  //
+  // A login method is a method which on success calls `this.setUserId(id)` and
+  // `Accounts._setLoginToken` on the server and returns an object with fields
+  // 'id' (containing the user id), 'token' (containing a resume token), and
+  // optionally `tokenExpires`.
+  //
+  // This function takes care of:
+  //   - Updating the Meteor.loggingIn() reactive data source
+  //   - Calling the method in 'wait' mode
+  //   - On success, saving the resume token to localStorage
+  //   - On success, calling Accounts.connection.setUserId()
+  //   - Setting up an onReconnect handler which logs in with
+  //     the resume token
+  //
+  // Options:
+  // - methodName: The method to call (default 'login')
+  // - methodArguments: The arguments for the method
+  // - validateResult: If provided, will be called with the result of the
+  //                 method. If it throws, the client will not be logged in (and
+  //                 its error will be passed to the callback).
+  // - userCallback: Will be called with no arguments once the user is fully
+  //                 logged in, or with the error on error.
+  //
+
+  callLoginMethod(options) {
+    var self = this;
+
+    options = _.extend({
+      methodName: 'login',
+      methodArguments: [{}],
+      _suppressLoggingIn: false
+    }, options);
+
+    // Set defaults for callback arguments to no-op functions; make sure we
+    // override falsey values too.
+    _.each(['validateResult', 'userCallback'], function (f) {
+      if (!options[f])
+        options[f] = function () {};
+    });
+
+    // Prepare callbacks: user provided and onLogin/onLoginFailure hooks.
+    var loginCallbacks = _.once(function (error) {
+      if (!error) {
+        self._onLoginHook.each(function (callback) {
+          callback();
+          return true;
+        });
+      } else {
+        self._onLoginFailureHook.each(function (callback) {
+          callback({ error });
+          return true;
+        });
+      }
+      options.userCallback.apply(this, arguments);
+    });
+
+    var reconnected = false;
+
+    // We want to set up onReconnect as soon as we get a result token back from
+    // the server, without having to wait for subscriptions to rerun. This is
+    // because if we disconnect and reconnect between getting the result and
+    // getting the results of subscription rerun, we WILL NOT re-send this
+    // method (because we never re-send methods whose results we've received)
+    // but we WILL call loggedInAndDataReadyCallback at "reconnect quiesce"
+    // time. This will lead to makeClientLoggedIn(result.id) even though we
+    // haven't actually sent a login method!
+    //
+    // But by making sure that we send this "resume" login in that case (and
+    // calling makeClientLoggedOut if it fails), we'll end up with an accurate
+    // client-side userId. (It's important that livedata_connection guarantees
+    // that the "reconnect quiesce"-time call to loggedInAndDataReadyCallback
+    // will occur before the callback from the resume login call.)
+    var onResultReceived = function (err, result) {
+      if (err || !result || !result.token) {
+        // Leave onReconnect alone if there was an error, so that if the user was
+        // already logged in they will still get logged in on reconnect.
+        // See issue #4970.
+      } else {
+        self.connection.onReconnect = function () {
+          reconnected = true;
+          // If our token was updated in storage, use the latest one.
+          var storedToken = self._storedLoginToken();
+          if (storedToken) {
+            result = {
+              token: storedToken,
+              tokenExpires: self._storedLoginTokenExpires()
+            };
+          }
+          if (! result.tokenExpires)
+            result.tokenExpires = self._tokenExpiration(new Date());
+          if (self._tokenExpiresSoon(result.tokenExpires)) {
+            self.makeClientLoggedOut();
+          } else {
+            self.callLoginMethod({
+              methodArguments: [{resume: result.token}],
+              // Reconnect quiescence ensures that the user doesn't see an
+              // intermediate state before the login method finishes. So we don't
+              // need to show a logging-in animation.
+              _suppressLoggingIn: true,
+              userCallback: function (error) {
+                var storedTokenNow = self._storedLoginToken();
+                if (error) {
+                  // If we had a login error AND the current stored token is the
+                  // one that we tried to log in with, then declare ourselves
+                  // logged out. If there's a token in storage but it's not the
+                  // token that we tried to log in with, we don't know anything
+                  // about whether that token is valid or not, so do nothing. The
+                  // periodic localStorage poll will decide if we are logged in or
+                  // out with this token, if it hasn't already. Of course, even
+                  // with this check, another tab could insert a new valid token
+                  // immediately before we clear localStorage here, which would
+                  // lead to both tabs being logged out, but by checking the token
+                  // in storage right now we hope to make that unlikely to happen.
+                  //
+                  // If there is no token in storage right now, we don't have to
+                  // do anything; whatever code removed the token from storage was
+                  // responsible for calling `makeClientLoggedOut()`, or the
+                  // periodic localStorage poll will call `makeClientLoggedOut`
+                  // eventually if another tab wiped the token from storage.
+                  if (storedTokenNow && storedTokenNow === result.token) {
+                    self.makeClientLoggedOut();
+                  }
+                }
+                // Possibly a weird callback to call, but better than nothing if
+                // there is a reconnect between "login result received" and "data
+                // ready".
+                loginCallbacks(error);
+              }});
+          }
+        };
+      }
+    };
+
+    // This callback is called once the local cache of the current-user
+    // subscription (and all subscriptions, in fact) are guaranteed to be up to
+    // date.
+    var loggedInAndDataReadyCallback = function (error, result) {
+      // If the login method returns its result but the connection is lost
+      // before the data is in the local cache, it'll set an onReconnect (see
+      // above). The onReconnect will try to log in using the token, and *it*
+      // will call userCallback via its own version of this
+      // loggedInAndDataReadyCallback. So we don't have to do anything here.
+      if (reconnected)
+        return;
+
+      // Note that we need to call this even if _suppressLoggingIn is true,
+      // because it could be matching a _setLoggingIn(true) from a
+      // half-completed pre-reconnect login method.
+      self._setLoggingIn(false);
+      if (error || !result) {
+        error = error || new Error(
+          "No result from call to " + options.methodName);
+        loginCallbacks(error);
+        return;
+      }
+      try {
+        options.validateResult(result);
+      } catch (e) {
+        loginCallbacks(e);
+        return;
+      }
+
+      // Make the client logged in. (The user data should already be loaded!)
+      self.makeClientLoggedIn(result.id, result.token, result.tokenExpires);
+      loginCallbacks();
+    };
+
+    if (!options._suppressLoggingIn)
+      self._setLoggingIn(true);
+    self.connection.apply(
+      options.methodName,
+      options.methodArguments,
+      {wait: true, onResultReceived: onResultReceived},
+      loggedInAndDataReadyCallback);
+  }
+
+  makeClientLoggedOut() {
+    // Ensure client was successfully logged in before running logout hooks.
+    if (this.connection._userId) {
+      this._onLogoutHook.each(function (callback) {
+        callback();
+        return true;
+      });
+    }
+    this._unstoreLoginToken();
+    this.connection.setUserId(null);
+    this.connection.onReconnect = null;
+  }
+
+  makeClientLoggedIn(userId, token, tokenExpires) {
+    this._storeLoginToken(userId, token, tokenExpires);
+    this.connection.setUserId(userId);
+  }
+
+
+  ///
+  /// LOGIN SERVICES
+  ///
+
+  // A reactive function returning whether the loginServiceConfiguration
+  // subscription is ready. Used by accounts-ui to hide the login button
+  // until we have all the configuration loaded
+  //
+  loginServicesConfigured() {
+    return this._loginServicesHandle.ready();
+  }
+
+
+  // Some login services such as the redirect login flow or the resume
+  // login handler can log the user in at page load time.  The
+  // Meteor.loginWithX functions have a callback argument, but the
+  // callback function instance won't be in memory any longer if the
+  // page was reloaded.  The `onPageLoadLogin` function allows a
+  // callback to be registered for the case where the login was
+  // initiated in a previous VM, and we now have the result of the login
+  // attempt in a new VM.
+
+  // Register a callback to be called if we have information about a
+  // login attempt at page load time.  Call the callback immediately if
+  // we already have the page load login attempt info, otherwise stash
+  // the callback to be called if and when we do get the attempt info.
+  //
+  onPageLoadLogin(f) {
+    if (this._pageLoadLoginAttemptInfo) {
+      f(this._pageLoadLoginAttemptInfo);
+    } else {
+      this._pageLoadLoginCallbacks.push(f);
+    }
+  }
+
+
+  // Receive the information about the login attempt at page load time.
+  // Call registered callbacks, and also record the info in case
+  // someone's callback hasn't been registered yet.
+  //
+  _pageLoadLogin(attemptInfo) {
+    if (this._pageLoadLoginAttemptInfo) {
+      Meteor._debug("Ignoring unexpected duplicate page load login attempt info");
+      return;
+    }
+
+    _.each(this._pageLoadLoginCallbacks, function (callback) {
+      callback(attemptInfo);
+    });
+
+    this._pageLoadLoginCallbacks = [];
+    this._pageLoadLoginAttemptInfo = attemptInfo;
+  }
+
+  ///
+  /// LOGIN TOKENS
+  ///
+
+  // Methods for storing a login token and user id in the
+  // browser's localStorage facility. It polls local storage every
+  // few seconds to synchronize login state between multiple tabs
+  // in the same browser.
+
+  loginWithToken(token, callback) {
+    this.callLoginMethod({
+      methodArguments: [{
+        resume: token
+      }],
+      userCallback: callback
+    });
+  }
+
+  // Semi-internal API. Call this function to re-enable auto login
+  // after if it was disabled at startup.
+  _enableAutoLogin() {
+    this._autoLoginEnabled = true;
+    this._pollStoredLoginToken();
+  }
+
+
+  ///
+  /// STORING
+  ///
+
+  // Call this from the top level of the test file for any test that does
+  // logging in and out, to protect multiple tabs running the same tests
+  // simultaneously from interfering with each others' localStorage.
+  _isolateLoginTokenForTest() {
+    this.LOGIN_TOKEN_KEY = this.LOGIN_TOKEN_KEY + Random.id();
+    this.USER_ID_KEY = this.USER_ID_KEY + Random.id();
+  }
+
+  _storeLoginToken(userId, token, tokenExpires) {
+    Meteor._localStorage.setItem(this.USER_ID_KEY, userId);
+    Meteor._localStorage.setItem(this.LOGIN_TOKEN_KEY, token);
+    if (! tokenExpires)
+      tokenExpires = this._tokenExpiration(new Date());
+    Meteor._localStorage.setItem(this.LOGIN_TOKEN_EXPIRES_KEY, tokenExpires);
+
+    // to ensure that the localstorage poller doesn't end up trying to
+    // connect a second time
+    this._lastLoginTokenWhenPolled = token;
+  }
+
+  _unstoreLoginToken() {
+    Meteor._localStorage.removeItem(this.USER_ID_KEY);
+    Meteor._localStorage.removeItem(this.LOGIN_TOKEN_KEY);
+    Meteor._localStorage.removeItem(this.LOGIN_TOKEN_EXPIRES_KEY);
+
+    // to ensure that the localstorage poller doesn't end up trying to
+    // connect a second time
+    this._lastLoginTokenWhenPolled = null;
+  }
+
+  // This is private, but it is exported for now because it is used by a
+  // test in accounts-password.
+  _storedLoginToken() {
+    return Meteor._localStorage.getItem(this.LOGIN_TOKEN_KEY);
+  }
+
+  _storedLoginTokenExpires() {
+    return Meteor._localStorage.getItem(this.LOGIN_TOKEN_EXPIRES_KEY);
+  }
+
+  _storedUserId() {
+    return Meteor._localStorage.getItem(this.USER_ID_KEY);
+  }
+
+  _unstoreLoginTokenIfExpiresSoon() {
+    const tokenExpires = this._storedLoginTokenExpires();
+    if (tokenExpires && this._tokenExpiresSoon(new Date(tokenExpires))) {
+      this._unstoreLoginToken();
+    }
+  }
+
+  ///
+  /// AUTO-LOGIN
+  ///
+
+  _initLocalStorage() {
+    // Key names to use in localStorage
+    this.LOGIN_TOKEN_KEY = "Meteor.loginToken";
+    this.LOGIN_TOKEN_EXPIRES_KEY = "Meteor.loginTokenExpires";
+    this.USER_ID_KEY = "Meteor.userId";
+
+    var rootUrlPathPrefix = __meteor_runtime_config__.ROOT_URL_PATH_PREFIX;
+    if (rootUrlPathPrefix || this.connection !== Meteor.connection) {
+      // We want to keep using the same keys for existing apps that do not
+      // set a custom ROOT_URL_PATH_PREFIX, so that most users will not have
+      // to log in again after an app updates to a version of Meteor that
+      // contains this code, but it's generally preferable to namespace the
+      // keys so that connections from distinct apps to distinct DDP URLs
+      // will be distinct in Meteor._localStorage.
+      var namespace = ":" + this.connection._stream.rawUrl;
+      if (rootUrlPathPrefix) {
+        namespace += ":" + rootUrlPathPrefix;
+      }
+      this.LOGIN_TOKEN_KEY += namespace;
+      this.LOGIN_TOKEN_EXPIRES_KEY += namespace;
+      this.USER_ID_KEY += namespace;
+    }
+
+    let token = undefined;
+    if (this._autoLoginEnabled) {
+      // Immediately try to log in via local storage, so that any DDP
+      // messages are sent after we have established our user account
+      this._unstoreLoginTokenIfExpiresSoon();
+      token = this._storedLoginToken();
+      if (token) {
+        // On startup, optimistically present us as logged in while the
+        // request is in flight. This reduces page flicker on startup.
+        const userId = this._storedUserId();
+        userId && this.connection.setUserId(userId);
+        this.loginWithToken(token, err => {
+          if (err) {
+            Meteor._debug("Error logging in with token: " + err);
+            this.makeClientLoggedOut();
+          }
+
+          this._pageLoadLogin({
+            type: "resume",
+            allowed: !err,
+            error: err,
+            methodName: "login",
+            // XXX This is duplicate code with loginWithToken, but
+            // loginWithToken can also be called at other times besides
+            // page load.
+            methodArguments: [{resume: token}]
+          });
+        });
+      }
+    }
+
+    // Poll local storage every 3 seconds to login if someone logged in in
+    // another tab
+    this._lastLoginTokenWhenPolled = token;
+
+    if (this._pollIntervalTimer) {
+      // Unlikely that _initLocalStorage will be called more than once for
+      // the same AccountsClient instance, but just in case...
+      clearInterval(this._pollIntervalTimer);
+    }
+
+    this._pollIntervalTimer = setInterval(() => {
+      this._pollStoredLoginToken();
+    }, 3000);
+  }
+
+  _pollStoredLoginToken() {
+    if (!this._autoLoginEnabled) {
+      return;
+    }
+
+    const currentLoginToken = this._storedLoginToken();
+
+    // != instead of !== just to make sure undefined and null are treated the same
+    if (this._lastLoginTokenWhenPolled != currentLoginToken) {
+      if (currentLoginToken) {
+        this.loginWithToken(currentLoginToken, err => {
+          if (err) {
+            this.makeClientLoggedOut();
+          }
+        });
+      } else {
+        this.logout();
+      }
+    }
+
+    this._lastLoginTokenWhenPolled = currentLoginToken;
+  }
 };
 
 var Ap = AccountsClient.prototype;
@@ -199,204 +630,6 @@ Meteor.loggingOut = function () {
   return Accounts.loggingOut();
 };
 
-///
-/// LOGIN METHODS
-///
-
-// Call a login method on the server.
-//
-// A login method is a method which on success calls `this.setUserId(id)` and
-// `Accounts._setLoginToken` on the server and returns an object with fields
-// 'id' (containing the user id), 'token' (containing a resume token), and
-// optionally `tokenExpires`.
-//
-// This function takes care of:
-//   - Updating the Meteor.loggingIn() reactive data source
-//   - Calling the method in 'wait' mode
-//   - On success, saving the resume token to localStorage
-//   - On success, calling Accounts.connection.setUserId()
-//   - Setting up an onReconnect handler which logs in with
-//     the resume token
-//
-// Options:
-// - methodName: The method to call (default 'login')
-// - methodArguments: The arguments for the method
-// - validateResult: If provided, will be called with the result of the
-//                 method. If it throws, the client will not be logged in (and
-//                 its error will be passed to the callback).
-// - userCallback: Will be called with no arguments once the user is fully
-//                 logged in, or with the error on error.
-//
-Ap.callLoginMethod = function (options) {
-  var self = this;
-
-  options = _.extend({
-    methodName: 'login',
-    methodArguments: [{}],
-    _suppressLoggingIn: false
-  }, options);
-
-  // Set defaults for callback arguments to no-op functions; make sure we
-  // override falsey values too.
-  _.each(['validateResult', 'userCallback'], function (f) {
-    if (!options[f])
-      options[f] = function () {};
-  });
-
-  // Prepare callbacks: user provided and onLogin/onLoginFailure hooks.
-  var loginCallbacks = _.once(function (error) {
-    if (!error) {
-      self._onLoginHook.each(function (callback) {
-        callback();
-        return true;
-      });
-    } else {
-      self._onLoginFailureHook.each(function (callback) {
-        callback({ error });
-        return true;
-      });
-    }
-    options.userCallback.apply(this, arguments);
-  });
-
-  var reconnected = false;
-
-  // We want to set up onReconnect as soon as we get a result token back from
-  // the server, without having to wait for subscriptions to rerun. This is
-  // because if we disconnect and reconnect between getting the result and
-  // getting the results of subscription rerun, we WILL NOT re-send this
-  // method (because we never re-send methods whose results we've received)
-  // but we WILL call loggedInAndDataReadyCallback at "reconnect quiesce"
-  // time. This will lead to makeClientLoggedIn(result.id) even though we
-  // haven't actually sent a login method!
-  //
-  // But by making sure that we send this "resume" login in that case (and
-  // calling makeClientLoggedOut if it fails), we'll end up with an accurate
-  // client-side userId. (It's important that livedata_connection guarantees
-  // that the "reconnect quiesce"-time call to loggedInAndDataReadyCallback
-  // will occur before the callback from the resume login call.)
-  var onResultReceived = function (err, result) {
-    if (err || !result || !result.token) {
-      // Leave onReconnect alone if there was an error, so that if the user was
-      // already logged in they will still get logged in on reconnect.
-      // See issue #4970.
-    } else {
-      self.connection.onReconnect = function () {
-        reconnected = true;
-        // If our token was updated in storage, use the latest one.
-        var storedToken = self._storedLoginToken();
-        if (storedToken) {
-          result = {
-            token: storedToken,
-            tokenExpires: self._storedLoginTokenExpires()
-          };
-        }
-        if (! result.tokenExpires)
-          result.tokenExpires = self._tokenExpiration(new Date());
-        if (self._tokenExpiresSoon(result.tokenExpires)) {
-          self.makeClientLoggedOut();
-        } else {
-          self.callLoginMethod({
-            methodArguments: [{resume: result.token}],
-            // Reconnect quiescence ensures that the user doesn't see an
-            // intermediate state before the login method finishes. So we don't
-            // need to show a logging-in animation.
-            _suppressLoggingIn: true,
-            userCallback: function (error) {
-              var storedTokenNow = self._storedLoginToken();
-              if (error) {
-                // If we had a login error AND the current stored token is the
-                // one that we tried to log in with, then declare ourselves
-                // logged out. If there's a token in storage but it's not the
-                // token that we tried to log in with, we don't know anything
-                // about whether that token is valid or not, so do nothing. The
-                // periodic localStorage poll will decide if we are logged in or
-                // out with this token, if it hasn't already. Of course, even
-                // with this check, another tab could insert a new valid token
-                // immediately before we clear localStorage here, which would
-                // lead to both tabs being logged out, but by checking the token
-                // in storage right now we hope to make that unlikely to happen.
-                //
-                // If there is no token in storage right now, we don't have to
-                // do anything; whatever code removed the token from storage was
-                // responsible for calling `makeClientLoggedOut()`, or the
-                // periodic localStorage poll will call `makeClientLoggedOut`
-                // eventually if another tab wiped the token from storage.
-                if (storedTokenNow && storedTokenNow === result.token) {
-                  self.makeClientLoggedOut();
-                }
-              }
-              // Possibly a weird callback to call, but better than nothing if
-              // there is a reconnect between "login result received" and "data
-              // ready".
-              loginCallbacks(error);
-            }});
-        }
-      };
-    }
-  };
-
-  // This callback is called once the local cache of the current-user
-  // subscription (and all subscriptions, in fact) are guaranteed to be up to
-  // date.
-  var loggedInAndDataReadyCallback = function (error, result) {
-    // If the login method returns its result but the connection is lost
-    // before the data is in the local cache, it'll set an onReconnect (see
-    // above). The onReconnect will try to log in using the token, and *it*
-    // will call userCallback via its own version of this
-    // loggedInAndDataReadyCallback. So we don't have to do anything here.
-    if (reconnected)
-      return;
-
-    // Note that we need to call this even if _suppressLoggingIn is true,
-    // because it could be matching a _setLoggingIn(true) from a
-    // half-completed pre-reconnect login method.
-    self._setLoggingIn(false);
-    if (error || !result) {
-      error = error || new Error(
-        "No result from call to " + options.methodName);
-      loginCallbacks(error);
-      return;
-    }
-    try {
-      options.validateResult(result);
-    } catch (e) {
-      loginCallbacks(e);
-      return;
-    }
-
-    // Make the client logged in. (The user data should already be loaded!)
-    self.makeClientLoggedIn(result.id, result.token, result.tokenExpires);
-    loginCallbacks();
-  };
-
-  if (!options._suppressLoggingIn)
-    self._setLoggingIn(true);
-  self.connection.apply(
-    options.methodName,
-    options.methodArguments,
-    {wait: true, onResultReceived: onResultReceived},
-    loggedInAndDataReadyCallback);
-};
-
-Ap.makeClientLoggedOut = function () {
-  // Ensure client was successfully logged in before running logout hooks.
-  if (this.connection._userId) {
-    this._onLogoutHook.each(function (callback) {
-      callback();
-      return true;
-    });
-  }
-  this._unstoreLoginToken();
-  this.connection.setUserId(null);
-  this.connection.onReconnect = null;
-};
-
-Ap.makeClientLoggedIn = function (userId, token, tokenExpires) {
-  this._storeLoginToken(userId, token, tokenExpires);
-  this.connection.setUserId(userId);
-};
-
 /**
  * @summary Log the user out.
  * @locus Client
@@ -417,61 +650,16 @@ Meteor.logoutOtherClients = function (callback) {
   return Accounts.logoutOtherClients(callback);
 };
 
-
-///
-/// LOGIN SERVICES
-///
-
-// A reactive function returning whether the loginServiceConfiguration
-// subscription is ready. Used by accounts-ui to hide the login button
-// until we have all the configuration loaded
-//
-Ap.loginServicesConfigured = function () {
-  return this._loginServicesHandle.ready();
+/**
+ * @summary Login with a Meteor access token.
+ * @locus Client
+ * @param {String} [token] The Meteor access token
+ * @param {Function} [callback] Optional callback. Called with no arguments on success, or with a single `Error` argument on failure.
+ * @importFromPackage meteor
+ */
+Meteor.loginWithToken = function(token, callback) {
+  return Accounts.loginWithToken(token, callback);
 };
-
-
-// Some login services such as the redirect login flow or the resume
-// login handler can log the user in at page load time.  The
-// Meteor.loginWithX functions have a callback argument, but the
-// callback function instance won't be in memory any longer if the
-// page was reloaded.  The `onPageLoadLogin` function allows a
-// callback to be registered for the case where the login was
-// initiated in a previous VM, and we now have the result of the login
-// attempt in a new VM.
-
-// Register a callback to be called if we have information about a
-// login attempt at page load time.  Call the callback immediately if
-// we already have the page load login attempt info, otherwise stash
-// the callback to be called if and when we do get the attempt info.
-//
-Ap.onPageLoadLogin = function (f) {
-  if (this._pageLoadLoginAttemptInfo) {
-    f(this._pageLoadLoginAttemptInfo);
-  } else {
-    this._pageLoadLoginCallbacks.push(f);
-  }
-};
-
-
-// Receive the information about the login attempt at page load time.
-// Call registered callbacks, and also record the info in case
-// someone's callback hasn't been registered yet.
-//
-Ap._pageLoadLogin = function (attemptInfo) {
-  if (this._pageLoadLoginAttemptInfo) {
-    Meteor._debug("Ignoring unexpected duplicate page load login attempt info");
-    return;
-  }
-
-  _.each(this._pageLoadLoginCallbacks, function (callback) {
-    callback(attemptInfo);
-  });
-
-  this._pageLoadLoginCallbacks = [];
-  this._pageLoadLoginAttemptInfo = attemptInfo;
-};
-
 
 ///
 /// HANDLEBARS HELPERS
