+++ packages/accounts-base/client_main.js
@@ -1,6 +1,5 @@
 import {AccountsClient} from "./accounts_client.js";
 import {AccountsTest} from "./url_client.js";
-import "./localstorage_token.js";
 
 /**
  * @namespace Accounts
